# Prioritize primary shard movement during shard allocation