--- a/server/src/main/java/org/opensearch/cluster/routing/RoutingNodes.java
+++ b/server/src/main/java/org/opensearch/cluster/routing/RoutingNodes.java
@@ -1361,6 +1361,10 @@ public class RoutingNodes implements Iterable<RoutingNode> {
                     queue.offer(iter);
                     return iter.next();
                 }
+
+                public void remove() {
+                    throw new UnsupportedOperationException();
+                }
             };
         }
     }