@jainankitk 2021-11-02
Benchmark results for shard relocation:

Before the change
```
# Run complete. Total time: 02:57:48

Benchmark                                                            (indicesShardsReplicasSourceTargetRecoveries)  Mode  Cnt    Score    Error  Units
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      2|        0|       1|      1|      1|  avgt   30    2.739 ±  0.078  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        0|       1|      1|      2|  avgt   30    2.701 ±  0.080  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        0|       1|      1|      5|  avgt   30    2.783 ±  0.082  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        0|       1|      1|     10|  avgt   30    2.946 ±  0.086  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        0|       1|      1|     10|  avgt   30    3.139 ±  0.116  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        0|       1|      1|     10|  avgt   30    3.996 ±  0.104  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      2|        0|      10|     10|      1|  avgt   30    2.802 ±  0.071  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        0|      10|      5|      2|  avgt   30    2.818 ±  0.086  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        0|      10|      5|      5|  avgt   30    3.007 ±  0.083  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        0|       5|     10|      5|  avgt   30    3.104 ±  0.080  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        0|      10|      5|      5|  avgt   30    3.332 ±  0.090  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        0|      10|     20|      6|  avgt   30    4.843 ±  0.146  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      1|        1|      10|     10|      1|  avgt   30    2.831 ±  0.075  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        1|      10|      3|      3|  avgt   30    2.807 ±  0.069  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        1|       5|     12|      5|  avgt   30    3.192 ±  0.103  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        1|      10|     10|      6|  avgt   30    3.708 ±  0.085  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        1|      10|      5|      8|  avgt   30    3.927 ±  0.107  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        1|       8|     17|      8|  avgt   30    5.873 ±  0.137  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      1|        2|      10|     10|      1|  avgt   30    2.891 ±  0.079  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      10|      5|      3|  avgt   30    3.039 ±  0.077  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        2|       5|     10|      5|  avgt   30    3.657 ±  0.103  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        2|      10|      8|      7|  avgt   30    4.465 ±  0.107  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        2|      13|     17|      5|  avgt   30    5.197 ±  0.180  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        2|      10|     20|      8|  avgt   30    8.709 ±  0.275  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      2|        1|      20|     20|      1|  avgt   30    3.026 ±  0.074  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        1|      20|     30|      1|  avgt   30    3.084 ±  0.093  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        1|      20|     10|      3|  avgt   30    3.244 ±  0.072  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        1|      20|      5|      5|  avgt   30    3.306 ±  0.102  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        1|      20|     23|      6|  avgt   30    5.350 ±  0.142  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        1|      40|     20|      8|  avgt   30    8.227 ±  0.315  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      50|     30|      1|  avgt   30    3.632 ±  0.108  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      50|     25|      1|  avgt   30    3.492 ±  0.096  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        1|      50|     33|      2|  avgt   30    4.184 ±  0.183  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        1|      40|     50|      2|  avgt   30    4.653 ±  0.133  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        1|      50|     70|      3|  avgt   30    7.001 ±  0.205  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        1|      60|     50|      3|  avgt   30    9.056 ±  0.314  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        2|      50|     50|      1|  avgt   30    4.539 ±  0.138  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      50|     30|      1|  avgt   30    3.611 ±  0.107  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        2|      50|     40|      2|  avgt   30    5.209 ±  0.165  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        2|      40|     50|      2|  avgt   30    5.591 ±  0.164  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        2|      50|     30|      6|  avgt   30    8.810 ±  0.307  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        2|      33|     55|      6|  avgt   30   12.814 ±  0.413  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         500|     60|       1|     100|    100|     12|  avgt   30  241.278 ±  9.901  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         500|     60|       1|     100|     40|     12|  avgt   30  210.525 ±  8.003  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         500|     60|       1|      40|    100|     12|  avgt   30  192.286 ±  8.678  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         50|      60|       1|     100|    100|      6|  avgt   30   33.909 ±  1.247  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         50|      60|       1|     100|     40|      6|  avgt   30   19.245 ±  0.744  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         50|      60|       1|      40|    100|      6|  avgt   30   18.052 ±  0.677  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      2|        0|       1|      1|      1|  avgt   30    2.840 ±  0.084  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        0|       1|      1|      2|  avgt   30    2.764 ±  0.070  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        0|       1|      1|      5|  avgt   30    3.002 ±  0.079  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        0|       1|      1|     10|  avgt   30    3.338 ±  0.085  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        0|       1|      1|     10|  avgt   30    3.833 ±  0.114  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        0|       1|      1|     10|  avgt   30    5.827 ±  0.191  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      2|        0|      10|     10|      1|  avgt   30    2.914 ±  0.072  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        0|      10|      5|      2|  avgt   30    2.936 ±  0.084  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        0|      10|      5|      5|  avgt   30    3.297 ±  0.090  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        0|       5|     10|      5|  avgt   30    3.596 ±  0.079  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        0|      10|      5|      5|  avgt   30    4.206 ±  0.116  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        0|      10|     20|      6|  avgt   30    7.453 ±  0.211  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      1|        1|      10|     10|      1|  avgt   30    2.966 ±  0.078  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        1|      10|      3|      3|  avgt   30    2.996 ±  0.095  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        1|       5|     12|      5|  avgt   30    3.597 ±  0.113  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        1|      10|     10|      6|  avgt   30    4.482 ±  0.156  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        1|      10|      5|      8|  avgt   30    5.260 ±  0.156  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        1|       8|     17|      8|  avgt   30   10.017 ±  0.313  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      1|        2|      10|     10|      1|  avgt   30    3.059 ±  0.077  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        2|      10|      5|      3|  avgt   30    3.217 ±  0.072  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        2|       5|     10|      5|  avgt   30    4.161 ±  0.110  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        2|      10|      8|      7|  avgt   30    5.320 ±  0.178  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        2|      13|     17|      5|  avgt   30    7.248 ±  0.276  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        2|      10|     20|      8|  avgt   30   14.381 ±  0.527  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      2|        1|      20|     20|      1|  avgt   30    3.263 ±  0.088  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        1|      20|     30|      1|  avgt   30    3.333 ±  0.072  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        1|      20|     10|      3|  avgt   30    3.716 ±  0.119  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        1|      20|      5|      5|  avgt   30    3.858 ±  0.104  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        1|      20|     23|      6|  avgt   30    7.026 ±  0.199  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        1|      40|     20|      8|  avgt   30   12.798 ±  0.400  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        2|      50|     30|      1|  avgt   30    4.012 ±  0.128  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        2|      50|     25|      1|  avgt   30    3.902 ±  0.147  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        1|      50|     33|      2|  avgt   30    4.776 ±  0.133  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        1|      40|     50|      2|  avgt   30    5.557 ±  0.131  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        1|      50|     70|      3|  avgt   30    8.994 ±  0.285  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        1|      60|     50|      3|  avgt   30   13.973 ±  0.465  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        2|      50|     50|      1|  avgt   30    5.255 ±  0.150  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|      3|        2|      50|     30|      1|  avgt   30    3.991 ±  0.103  ms/op
AllocationBenchmark.measureShardRelocationComplete                  10|     10|        2|      50|     40|      2|  avgt   30    5.965 ±  0.210  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      1|        2|      40|     50|      2|  avgt   30    6.508 ±  0.223  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|      3|        2|      50|     30|      6|  avgt   30   11.108 ±  0.412  ms/op
AllocationBenchmark.measureShardRelocationComplete                 100|     10|        2|      33|     55|      6|  avgt   30   20.016 ±  0.746  ms/op
AllocationBenchmark.measureShardRelocationComplete                  500|     60|       1|     100|    100|     12|  avgt   30  540.946 ± 20.389  ms/op
AllocationBenchmark.measureShardRelocationComplete                  500|     60|       1|     100|     40|     12|  avgt   30  504.320 ± 19.405  ms/op
AllocationBenchmark.measureShardRelocationComplete                  500|     60|       1|      40|    100|     12|  avgt   30  461.199 ± 19.855  ms/op
AllocationBenchmark.measureShardRelocationComplete                  50|      60|       1|     100|    100|      6|  avgt   30   50.915 ±  2.835  ms/op
AllocationBenchmark.measureShardRelocationComplete                  50|      60|       1|     100|     40|      6|  avgt   30   33.995 ±  1.205  ms/op
AllocationBenchmark.measureShardRelocationComplete                  50|      60|       1|      40|    100|      6|  avgt   30   32.178 ±  1.052  ms/op
```

After the change
```
# Run complete. Total time: 02:58:39

Benchmark                                                            (indicesShardsReplicasSourceTargetRecoveries)  Mode  Cnt    Score    Error  Units
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      2|        0|       1|      1|      1|  avgt   30    2.782 ±  0.082  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        0|       1|      1|      2|  avgt   30    2.689 ±  0.085  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        0|       1|      1|      5|  avgt   30    2.800 ±  0.076  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        0|       1|      1|     10|  avgt   30    2.961 ±  0.075  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        0|       1|      1|     10|  avgt   30    3.128 ±  0.088  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        0|       1|      1|     10|  avgt   30    4.044 ±  0.112  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      2|        0|      10|     10|      1|  avgt   30    2.790 ±  0.072  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        0|      10|      5|      2|  avgt   30    2.804 ±  0.082  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        0|      10|      5|      5|  avgt   30    3.024 ±  0.107  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        0|       5|     10|      5|  avgt   30    3.095 ±  0.076  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        0|      10|      5|      5|  avgt   30    3.416 ±  0.100  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        0|      10|     20|      6|  avgt   30    4.947 ±  0.150  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      1|        1|      10|     10|      1|  avgt   30    2.825 ±  0.072  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        1|      10|      3|      3|  avgt   30    2.823 ±  0.076  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        1|       5|     12|      5|  avgt   30    3.140 ±  0.074  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        1|      10|     10|      6|  avgt   30    3.620 ±  0.084  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        1|      10|      5|      8|  avgt   30    3.976 ±  0.121  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        1|       8|     17|      8|  avgt   30    6.216 ±  0.189  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      1|        2|      10|     10|      1|  avgt   30    2.824 ±  0.071  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      10|      5|      3|  avgt   30    3.042 ±  0.089  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        2|       5|     10|      5|  avgt   30    3.673 ±  0.150  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        2|      10|      8|      7|  avgt   30    4.581 ±  0.121  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        2|      13|     17|      5|  avgt   30    5.330 ±  0.180  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        2|      10|     20|      8|  avgt   30    8.737 ±  0.288  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      2|        1|      20|     20|      1|  avgt   30    2.995 ±  0.069  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        1|      20|     30|      1|  avgt   30    3.086 ±  0.082  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|     10|        1|      20|     10|      3|  avgt   30    3.280 ±  0.089  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      1|        1|      20|      5|      5|  avgt   30    3.350 ±  0.162  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|      3|        1|      20|     23|      6|  avgt   30    5.380 ±  0.162  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard        100|     10|        1|      40|     20|      8|  avgt   30    8.410 ±  0.294  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      50|     30|      1|  avgt   30    3.552 ±  0.119  ms/op
AllocationBenchmark.measureExclusionOnZoneAwareStartedShard         10|      3|        2|      50|     25|      1|  avgt   30    3.413 ±  0.101  ms/op
…